+++ src/dashboardWebView/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './MessageHandlers';
 export * from './getRelPath';
 export * from './preserveColor';
 export * from './updateCssVariables';